requirements.txt
pip==25.3
requests==2.32.5
pytelegrambotapi==4.29.1
PyYAML==6.0.3
pydantic==2.12.5
pytest==9.0.2
tzdata==2025.3
yandex-music==2.2.0